+++ src/main/java/org/example/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.pieces;
 
 public class Pawn {
